enable all warnings (-Weverything) and fix many of them
• mostly signedness and unused variable problems
• careful with LLSP: we count some loops backwards with an index that is conceptually unsigned, but the check >= 0 requires signed
• also preprocessor #warning is non-standard (whoops)
• treat SDL and FFmpeg as system headers so we get no warnings there
• ffplay really does not like -Weverything

+++ Experiments/02 Times and Metrics/Components/metrics.c
@@ -34,7 +34,7 @@ void hook_slice_any(const AVCodecContext *c)
 		gettimeofday(&tv, NULL);
 		double usecs_stop = 1000000.0 * tv.tv_sec + (double)tv.tv_usec;
 		
-		usecs_per_tsc = (usecs_stop - usecs_start) / (tsc_stop - tsc_start);
+		usecs_per_tsc = (float)((usecs_stop - usecs_start) / (tsc_stop - tsc_start));
 	});
 	
 	printf("%d, %d, %d, %d, %d, %d, %d, %d, %d, %d, %d, %d, %d, %d, ",
